server: Fix rtsp_connection::parse segfault
Cleanups split_line, implements memsep as a replacement to strsep, and
replaces calls to strchr with memchr.

Can't replace strstr with strnstr because Ubuntu Precise's libbsd doesn't
implement it.

Other functions should be fixed to take into account the buffer limits
instead of relying on the string termination.

+++ server/rtsp.h
@@ -83,7 +83,7 @@ private:
 
 	std::map<int,rtsp_session*> sessions;
 
-	int parse();
+	ssize_t parse();
 	void sendResponse(const rtsp_message &response);
 
 	bool authenticate(rtsp_message &request, int device_id = -1);
